leverage mocha features to set up env for server tests
