+++ test/specs/responses.test.js
@@ -1,7 +1,7 @@
 const chai = require('chai');
 const expect = chai.expect;
 
-module.exports = (fetchMock) => {
+const {fetchMock} = testGlobals;
 	describe('responses', () => {
 		let fm;
 		before(() => {
@@ -385,4 +385,3 @@ module.exports = (fetchMock) => {
 			});
 		});
 	});
-};
